### 2023-02-09 #5139 @editorialbot
[REVIEW]: VisWaterNet: A Python package for visualization of water distribution networks
**Submitting author:** @meghnathomas (<a href="http://orcid.org/0000-0002-3755-9188">Meghna Thomas</a>)
**Repository:** https://github.com/tylertrimble/viswaternet
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0a
