### 2023-02-21 tylertrimble/viswaternet#101 @thurber
[JOSS Review] Feedback for JOSS review
Hello! I'll be collecting my feedback for the JOSS review in this issue. Feel free to create individual issues from each item if it works better for you. First of all, nice work! This is a really neat and useful way to visualize river networks and network data! I'll start by reviewing the administrative stuff, followed by the repo/documentation, and then finally the paper itself. Don't feel pressured to change everything I recommend; you can refute any point! For context, I'm working through the review checklist found in the JOSS review issue here: https://github.com/openjournals/joss-reviews/issues/5139.

[General]
- Can you comment on the contributions of the paper co-author Lina Sela? I only ask because they don't show up in the commit history, which is fine, just need to clarify.
- Can you add a section to the `README.md` or a `contributing.md` file outlining your community guidelines (i.e. how folks can/should contribute to the code). See GitHub's recommendations [here](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors).
- Can you add a section to the `README.md` that says how to run the automated tests
- There are a few deprecation warnings generated when running the tests; in particular the matplotlib ones may be cause for concern -- you don't want upstream changes to matplotlib to break your code. It would be preferable to update your use of matplotlib APIs to the latest syntax, but you could alternatively pin the maximum allowed matplotlib version in your `requirements.txt`.

[Docs/Code]
- In the GitHub README.md example (also in the Getting Started docs section), in case the user is not running in a notebook it would be helpful to explicitly mention the need to run `plt.show()` to see the plot.
- When running through the example in the GitHub README.md, a warning is generated -- would be good to either suppress it if it's irrelevant to your package, or more fully explain it so the user knows what should be done: `/Users/thur961/viswaternet/venv/lib/python3.10/site-packages/wntr/epanet/io.py:2082: UserWarning: Not all curves were used in "/Users/thur961/viswaternet/CTown.inp"; added with type None, units conversion left to user warnings.warn('Not all curves were used in "{}"; added with type None, units conversion left to user'.format(self.wn.name))`
- In Example 2, when plotting the pumps as dotted lines, it would be nice if the legend also updated to show pumps in dotted lines. Furthermore, it would be good to point out that the arguments are following the matplotlib notation for colors and line types, etc., and perhaps link to the matplotlib docs page that describes them.
- For all plots, it would be good to show the units along with the legend, and to perhaps somehow annotate the chosen variable and timestep or temporal aggregation method (I see that user can input a custom legend title as one option; if that's preferred then would help to mention it in the docs).
- Could be nice to support an `ax=None` default argument that creates a matplotlib fig/ax on behalf of the user. This would follow the precedent of libraries like `pandas`/`xarray`.

[Paper]
- to do
