### 2023-02-27 a8b8724
Create CONTRIBUTING.rst
