### 2023-02-27 48ce37e
fixed #102 contributing + plt.show()
(1) added link to CONTRIBUTING.rst section
(2) added note regarding use of plt.show()
