### 2023-02-27 e2b70ed
added link to automated test instructions in docs #102
