### 2023-02-28 tylertrimble/viswaternet#111 @erexer
JOSS review feedback
Hi! I'm one of the JOSS reviewers mentioned over in https://github.com/openjournals/joss-reviews/issues/5139. Super cool software and great work!

- Contribution: how was tylertrimble involved? The only info I have is the commit history so I want to double check that the authors are "appropriate and complete"
    - [https://github.com/tylertrimble/viswaternet/commits?author=tylertrimble&since=2023-02-01&until=2023-02-25](https://github.com/tylertrimble/viswaternet/commits?author=tylertrimble&since=2023-02-01&until=2023-02-25)
- Installation:
    - `pip install viswaternet` failed with python 3.11.0
        - `ERROR: No matching distribution found for numpy==1.21.4`
        - since you list as working with python 3.8+, it should work for 3.11.0. Let me know if you’re able to reproduce this, if not I can give a better trace
    - python 3.10.9 `python -m unittest` throws a lot of warnings
        - `MatplotlibDeprecationWarning: The get_cmap function was deprecated in Matplotlib 3.7 and will be removed two minor releases later. Use ``matplotlib.colormaps[name]`` or ``matplotlib.colormaps.get_cmap(obj)`` instead. cmap = mpl.cm.get_cmap(cmap)`
- Examples:
    - They look great! I’d recommend standardizing the example numbering for the python scripts and the examples in the docs. `example2_network_diameters.py` would be named `example8_network_diameters.py` and produce identical output to [Example 8](https://viswaternet.readthedocs.io/en/latest/examples.html#example-8-categorical-data-plot-for-link-diameter) in the readthedocs. (personal preference so feel free to ignore)
    - Getting the same error from every example:
        - `UserWarning: Not all curves were used in "Networks/CTown.inp"; added with type None, units conversion left to user warnings.warn('Not all curves were used in "{}"; added with type None, units conversion left to user'.format([self.wn.name](http://self.wn.name/)))`
- Tests
    - python 3.10.9 throws warning from `python -m unittest`
        - `MatplotlibDeprecationWarning: Passing *transOffset* without *offsets* has no effect. This behavior is deprecated since 3.5 and in 3.6, *transOffset* will begin having an effect regardless of *offsets*. In the meantime, if you wish to set *transOffset*, call collection.set_offset_transform(transOffset) explicitly.`
        - still runs the tests fine
- Documentation
    - would recommend adding the documentation link in the GitHub sidebar
    - also recommend running it through a spell checker since there were a few spelling mistakes
- Paper
    - Does “extracts WDS characteristics and simulation results of a user-provided .INP file” mean something different than “1. Import and visualize custom data generated either in Python scripts or in Excel files.” are the Python scripts or Excel files generating anything other than .INP files?
    - recommend changing “Add a second dimension of information to a plot by customizing node markers and sizes and link styles and widths.” to “Add … customizing node markers/sizes and link styles/widths.” for clarity.
    - “Convert units (e.g., a user can choose to represent length data in units of *m*, *cm*, *ft*, or *in*.” missing closing parenthesis.
