### 2023-03-03 9cc0608
Address #108 and add two new arguments to draw_color_bar
