### 2023-03-05 64a405f
Added new function to generate legend/colorbar titles
