+++ viswaternet/utils/__init__.py
@@ -4,3 +4,4 @@ from .normalize_parameter import normalize_parameter
 from .save_fig import save_fig
 from .unit_conversion import unit_conversion
 from .fancyarrowpatch_to_linecollection import fancyarrowpatch_to_linecollection
+from .label_generator import label_generator
