### 2023-03-05 26bc84e
Changes legend patches to lines and changes when elements are drawn
Addresses Issue #104
