+++ viswaternet/drawing/unique.py
@@ -398,6 +398,9 @@ def plot_unique_data(
         Determines if color bar is drawn.
     """
     model = self.model
+    
+    if len(self.model['G_list_pumps_only'])==0:
+        pumps = False
     if parameter == "demand_patterns":
 
         demand_pattern_nodes, patterns = processing.get_demand_patterns(self)
@@ -464,7 +467,12 @@ def plot_unique_data(
                 legend_title_font_size=legend_title_font_size,
                 draw_frame=draw_frame,
                 pump_color=pump_color,
-                base_link_color=base_link_color
+                base_link_color=base_link_color,
+                pump_line_style=pump_line_style,
+                base_link_line_style=base_link_line_style,
+                base_link_arrows=base_link_arrows,
+                pump_arrows=pump_arrows,
+                draw_base_links=True,
             )
         if savefig:
 
@@ -564,7 +572,12 @@ def plot_unique_data(
                 legend_title_font_size=legend_title_font_size,
                 draw_frame=draw_frame,
                 pump_color=pump_color,
-                base_link_color=base_link_color
+                base_link_color=base_link_color,
+                pump_line_style=pump_line_style,
+                base_link_line_style=base_link_line_style,
+                base_link_arrows=base_link_arrows,
+                pump_arrows=pump_arrows,
+                draw_base_links=False,
             )
         if savefig:
 
@@ -672,7 +685,12 @@ def plot_unique_data(
                 legend_title_font_size=legend_title_font_size,
                 draw_frame=draw_frame,
                 pump_color=pump_color,
-                base_link_color=base_link_color
+                base_link_color=base_link_color,
+                pump_line_style=pump_line_style,
+                base_link_line_style=base_link_line_style,
+                base_link_arrows=base_link_arrows,
+                pump_arrows=pump_arrows,
+                draw_base_links=True,
             )
         if savefig:
 
@@ -733,6 +751,28 @@ def plot_unique_data(
                     base_link_line_style=base_link_line_style,
                     base_link_arrows=base_link_arrows
                 )
+                
+                if legend:
+
+                    base.draw_legend(
+                        ax,
+                        custom_data_values[1],
+                        title=legend_title,
+                        pumps=pumps,
+                        loc=legend_loc_1,
+                        loc2=legend_loc_2,
+                        font_size=font_size,
+                        font_color=font_color,
+                        legend_title_font_size=legend_title_font_size,
+                        draw_frame=draw_frame,
+                        pump_color=pump_color,
+                        base_link_color=base_link_color,
+                        pump_line_style=pump_line_style,
+                        base_link_line_style=base_link_line_style,
+                        base_link_arrows=base_link_arrows,
+                        pump_arrows=pump_arrows,
+                        draw_base_links=False
+                    )
             elif parameter_type == "node":
 
                 discrete.draw_discrete_nodes(
@@ -782,23 +822,32 @@ def plot_unique_data(
                     base_link_width=base_link_width,
                     base_link_line_style=base_link_line_style,
                     base_link_arrows=base_link_arrows
+                
+                
                 )
-            if legend:
+                
+                if legend:
 
-                base.draw_legend(
-                    ax,
-                    custom_data_values[1],
-                    title=legend_title,
-                    pumps=pumps,
-                    loc=legend_loc_1,
-                    loc2=legend_loc_2,
-                    font_size=font_size,
-                    font_color=font_color,
-                    legend_title_font_size=legend_title_font_size,
-                    draw_frame=draw_frame,
-                    pump_color=pump_color,
-                    base_link_color=base_link_color
-                )
+                    base.draw_legend(
+                        ax,
+                        custom_data_values[1],
+                        title=legend_title,
+                        pumps=pumps,
[…]
