### 2023-03-05 11924fd
Add default axis creation if none is provided
Addresses issue #106
