### 2023-03-05 ca61f31
Update example files to utilize new default axis creation
