+++ Examples/example2_network_diameters.py
@@ -10,15 +10,9 @@ import numpy as np
 # Initialize VisWaterNet model
 model = vis.VisWNModel(r"Networks/CTown.inp")  
 
-# Initialize plot
-fig, ax = plt.subplots(figsize=(12, 12))  
-
-# Hide frame on plot
-ax.set_frame_on(False)
-
 # Plot unique pipe diameters
 model.plot_unique_data(
-    ax, parameter="diameter", 
+    parameter="diameter", 
     unit="in", 
     interval_link_width_list=np.linspace(1,7,10),
     cmap="Blues", # Controls link widths
@@ -27,5 +21,6 @@ model.plot_unique_data(
     legend_sig_figs=0, # Whole numbers only
     pump_color="mediumseagreen", # Customize pump color and width
     pump_width=5) # In Pixels 
-    
+
+#Called in the case that interactive plotting isn't enabled
 plt.show()
