+++ Examples/example5_nodalerror.py
@@ -14,18 +14,12 @@ from example5_nodalerror_generation import wn,error_list,element_list
 # Initialize VisWaterNet model
 model = vis.VisWNModel(r"Networks/CTown.inp", network_model=wn)
 
-# Initialize plot
-fig, ax = plt.subplots(figsize=(12, 12))  
-
-# Hide frame on plot
-ax.set_frame_on(False)
-
 # Plot custom data 
 model.plot_unique_data(
-    ax,
     parameter="custom_data",node_size=200,line_widths=1,
     edge_colors="k",parameter_type="node",data_type="continuous",
     custom_data_values=[element_list, error_list],
     color_bar_title="Error (%)",cmap="bwr",draw_color_bar=True)
 
+#Called in the case that interactive plotting isn't enabled
 plt.show()
