### 2023-03-05 5262099
Replace get_cmap with colormaps[cmap]
