### 2023-03-05 49f7072
Adds missing case in label_generator
