+++ .github/workflows/python-app.yml
@@ -19,7 +19,7 @@ jobs:
     strategy:
       matrix:
         python-version: ['3.8', '3.9', '3.10']
-        os: [windows-latest, macOS-latest, ubuntu-latest] 
+        os: [ubuntu-latest] 
     steps:
     - uses: actions/checkout@v2
     - name: Set up Python
