### 2023-03-05 24dcb04
document common warnings in docs close #107
