### 2023-03-05 a7d87c6
add warnings to docs
