### 2023-03-05 4e55029
update paper based on comments from #114
