+++ docs/examples.rst
@@ -23,7 +23,7 @@ Alternatively, we can initialize a VisWaterNet model corresponding to a `WNTR`_ 
 
 .. _`WNTR`: https://www.osti.gov/biblio/1376816
 
-3. Initialize and customize a matplotlib figure and axis.
+3. Initialize and customize a Matplotlib figure and axis.
 
 .. code:: python
 
@@ -96,7 +96,7 @@ Here, we create a **continuous** data plot for mean **link flow rate** over the 
    :width: 600
    :alt: Continuous link plot
 
-Next we demonstrate how to visualize data in a discete manner, i.e., by grouping data into intervals and assigning colors according to each interval shown in a legend.
+Next we demonstrate how to visualize data in a discrete manner, i.e., by grouping data into intervals and assigning colors according to each interval shown in a legend.
 
 Example 5 - Discrete Data Plot for Nodal Demand
 -----------------------------
@@ -107,7 +107,8 @@ Here, we create a **discrete** data plot for **nodal demand at hour 10**. We spe
 
 .. code:: python
 
-    model.plot_discrete_nodes(ax,parameter = "demand", value = 10, num_intervals = 3, legend_loc_2 = 'upper left', unit = 'CMH')
+    model.plot_discrete_nodes(ax, parameter = "demand", value = 10, num_intervals = 3, 
+                              legend_loc_2 = 'upper left', unit = 'CMH')
     
 .. _basic5:
 .. figure:: figures/eg_plot_5.png
@@ -117,12 +118,13 @@ Here, we create a **discrete** data plot for **nodal demand at hour 10**. We spe
 Example 6 - Discrete Data Plot for Link Velocity
 -----------------------------
 
-Here, we create a **discrete** data plot for **maximum link velocity** over the simulation duration. We specify the intervals we would like to see (*0-2, 2-6, 6-10*) and VisWaterNet groups all parameter data into these intervals, and constructs extra intervals (here, *<0* or *>10*) if any data points fall outside of the specified intervals. We customize the legend by specifying that the legend labels should have zero digits after the decimal point (*legend_sig_figs=0*) and providing a legend title. We also convert the units of velocity to ft/s (from the default SI units of m/s). 
+Here, we create a **discrete** data plot for **maximum link velocity** over the simulation duration. We specify the intervals we would like to see (*0-2, 2-6, 6-10*). VisWaterNet groups all parameter data into these intervals, and constructs extra intervals (here, *<0* or *>10*) if any data points fall outside of the specified intervals. We customize the legend by specifying that the legend labels should have zero digits after the decimal point (*legend_sig_figs=0*) and providing a legend title. We also convert the units of velocity to ft/s (from the default SI units of m/s). 
 
 .. code:: python
 
-    model.plot_discrete_links(ax,parameter = "velocity", value = 'max', intervals = [0,2,6,10], legend_title = 'Link velocity [ft/s]', 
-                          legend_sig_figs = 0, legend_loc_2 = 'lower left', unit = 'ft/s')
+    model.plot_discrete_links(ax,parameter = "velocity", value = 'max', intervals = [0,2,6,10], 
+                              legend_title = 'Link velocity [ft/s]', legend_sig_figs = 0, 
+                              legend_loc_2 = 'lower left', unit = 'ft/s')
 
 .. _basic6:
 .. figure:: figures/eg_plot_6.png
@@ -130,45 +132,46 @@ Here, we create a **discrete** data plot for **maximum link velocity** over the 
    :alt: Continuous link plot
    
 Next, we demonstrate the different functionalities offered by the ``plot_unique data`` function:
-1. visualizing categorical data, i.e., specific properties of nodes or links are represented as a fixed set of categories 
[…]
