### 2023-03-07 6a49aa8
update python-app.yml to test on windows and mac
