+++ .github/workflows/python-app.yml
@@ -19,7 +19,7 @@ jobs:
     strategy:
       matrix:
         python-version: ['3.8', '3.9', '3.10']
-        os: [ubuntu-latest] 
+        os: [windows-latest, macOS-latest,ubuntu-latest] 
     steps:
     - uses: actions/checkout@v2
     - name: Set up Python
@@ -30,7 +30,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install flake8 pytest
-        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+        pip install -r requirements.txt
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
