### 2023-03-07 8d62736
place warnings in contributing file
