### 2023-03-07 b539fde
update index.rst -- delete warnings
