### 2023-03-07 f6ea50e
double backslash #115
seeing if adding a second backslash resolves the warning we get in the tests based on http://www.pythondoc.com/sphinx/ext/math.html
