### 2023-03-07 d329ab2
Change wn.get_graph() to wn.to_graph()
