+++ docs/source/modules.rst
@@ -1,4 +1,4 @@
-API documentation
+API Documentation
 =======
 
 .. toctree::
