+++ .github/workflows/python-app.yml
@@ -12,10 +12,6 @@ on:
 permissions:
   contents: read
 
-runs:
-  using: 'node16'
-  main: 'main.js'
-
 jobs:
   build:
 
@@ -30,6 +26,7 @@ jobs:
       uses: actions/setup-python@v4
       with:
         python-version: ${{ matrix.python-version }}
+        node-version: 16
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
