+++ .github/workflows/python-app.yml
@@ -21,12 +21,11 @@ jobs:
         python-version: ['3.8', '3.9', '3.10']
         os: [windows-latest, macOS-latest,ubuntu-latest] 
     steps:
-    - uses: actions/checkout@v2
+    - uses: actions/checkout@v3
     - name: Set up Python
       uses: actions/setup-python@v4
       with:
         python-version: ${{ matrix.python-version }}
-        node-version: 16
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
