### 2023-03-10 9f6edfb
updated examples to close #110
added new files to correspond to Example Applications section of docs
renamed files
updated scripts to correspond to recent code updates
added figures in separate folder so users know what the resulting figs look like
