### 2023-03-10 1ff728c
update label generator
change units to common EPANET form
add velocity to parameter list
