### 2023-03-10 3dddbe0
update draw_base_elements
previously if we tried to make subplots of basic layout plots using plot_basic_elements, the nodes of all models would show up only on the far right plot
added ax parameter to draw_base_elements so nodes are drawn in the correct axis
