### 2023-03-14 2caeaa5
change back to wn.get_graph() from wn.to_graph()
