### 2023-03-14 a271220
set lower and upper bounds on wntr version
