+++ .github/workflows/python-app.yml
@@ -18,7 +18,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        python-version: ['3.7', '3.8', '3.9', '3.10']
+        python-version: ['3.8', '3.9', '3.10']
         os: [windows-latest, macOS-latest,ubuntu-latest] 
     steps:
     - uses: actions/checkout@v3
