+++ requirements.txt
@@ -1,6 +1,6 @@
 pandas
 numpy
 matplotlib>=3.5.0
-wntr>=0.5.0
+wntr<0.5.0
 imageio
 networkx>=2.7
