### 2023-03-15 558a894
remove color bar
