+++ examples/example10_nodalerror.py
@@ -17,5 +17,5 @@ model.plot_unique_data(parameter = "custom_data", node_size = 200,
                        parameter_type = "node", data_type = "continuous",    
                        custom_data_values = [element_list, error_list],
                        color_bar_title = "Error (%)", cmap = "bwr",
-                       draw_color_bar=True, save_name = 'figures/example10', dpi = 400)
+                       save_name = 'figures/example10', dpi = 400)
 plt.show()
