### 2023-03-15 911199d
reflect code updates
updated figures
updated code snippets
updated instructions (ax = None/ax)
