+++ docs/examples.rst
@@ -2,14 +2,13 @@
 Example Applications
 ====================
 
-All VisWaterNet scripts should begin with the following three steps:
+All VisWaterNet scripts should begin with the following steps:
 
-1.  Import the VisWaterNet and matplotlib.pyplot packages.
+1. Import the VisWaterNet package.
 
 .. code:: python
 
     import viswaternet as vis
-    import matplotlib.pyplot as plt
     
 2. Initialize a VisWaterNet model object for the .INP file of the water distribution network. For these examples, we use the CTown network model introduced by `Ostfeld (2016)`_.
 
@@ -17,20 +16,22 @@ All VisWaterNet scripts should begin with the following three steps:
 
 .. code:: python
 
-    model = vis.VisWNModel("CTown.inp")
+    model = vis.VisWNModel('Networks/CTown.inp')
     
 Alternatively, we can initialize a VisWaterNet model corresponding to a `WNTR`_ water network model object.
 
 .. _`WNTR`: https://www.osti.gov/biblio/1376816
 
-3. Initialize and customize a Matplotlib figure and axis.
+3. If you would like to draw the plot into a figure you want to customize yourself (e.g., by drawing into a subplot axis, or by choosing the height and width), import the Matplotlib package and initialize a Matplotlib figure and axis. This is optional; by default, if not provided an empty axis or figure, VisWaterNet will create a figure on behalf of the user.
 
 .. code:: python
 
-    fig, ax = plt.subplots(figsize=(11,11))  
-    ax.set_frame_on(False) # remove frame from figure
+    import matplotlib.pyplot as plt
     
-After we have initialized our VisWaterNet model object and empty matplotlib figure, we can proceed to call on different functions offered by the VisWaterNet library to generate a variety of figures. Below, we provide a series of examples to highlight the different VisWaterNet plotting functions and their wide range of inputs.
+    fig, ax = plt.subplots(figsize=(11,11))  
+    ax.set_frame_on(False)
+    
+After we have initialized our VisWaterNet model object, we can proceed to call on different functions offered by the VisWaterNet library to generate a variety of figures. Below, we provide a series of examples to highlight the different VisWaterNet plotting functions and their wide range of inputs.
 
 Example 1 - Basic Network Layout Plot
 -----------------------------
@@ -39,25 +40,27 @@ This example demonstrates the basic plotting functionality provided by VisWaterN
 
 .. code:: python
 
-    model.plot_basic_elements(ax)
+    model.plot_basic_elements()
 
 .. _basic1:
-.. figure:: figures/eg_plot_1.png
+.. figure:: figures/example1.png
    :width: 600
    :alt: Basic network layout
 
 Example 2 - Customizing a Basic Network Layout Plot
 -----------------------------
 
-Here, we customize the **basic** network plot by changing the location of the legend, color of the tank marker, and pump line style.
+Here, we customize the **basic** network plot by changing the location of the legend, color of the tank marker, and pump line style, and draw the figure into axis *ax*.
 
 .. code:: python
 
-    model.plot_basic_elements(ax, legend_loc = 'upper left', 
-                              tank_color = 'g', pump_line_style = ':')
+    model.plot_basic_elements(ax, 
+                              legend_loc = 'lower left', 
+                              tank_color = 'g', 
+                              pump_line_style = ':')
 
 .. _basic2:
-.. figure:: figures/eg_plot_2.png
+.. figure:: figures/example2.png
    :width: 600
    :alt: Basic network layout modified
 
@@ -71,14 +74,16 @@ Next, Examples 3 and 4 demonstrate how to visualize data in a continuous manner,
 Example 3 - Continuous Node Data Plot for Nodal Pressure
 -----------------------------
  
[…]
