### 2023-03-15 aad7484
add figs to readme
