### 2023-03-15 efb35d3
Update AUTHORS.rst
