+++ AUTHORS.rst
@@ -5,10 +5,10 @@ Credits
 Development Lead
 ----------------
 
-* Tyler Trimble <TylerL.Trimble@gmail.com>
-* Meghna Thomas <meghnathomas@utexas.edu>
-* Lina Sela <linasela@utexas.edu>
+* Tyler Trimble <TylerL.Trimble@gmail.com>: Software, Methodology, Validation 
+* Meghna Thomas <meghnathomas@utexas.edu>: Software, Methodology, Writing - Original Draft
+* Lina Sela <linasela@utexas.edu>: Conceptualization, Supervision, Funding acquisition, Writing - Review & Editing
 
 Contributors
 ------------
-None yet. Why not be the first?
+This package is being developed and supported by the WATSUP group at UT Austin.
