+++ viswaternet/drawing/base.py
@@ -29,6 +29,7 @@ def draw_nodes(
     label=None,
 ):
     """Draws continuous nodal data onto the figure.
+    
     Arguments
     ---------
     ax : axes._subplots.AxesSubplot
@@ -46,11 +47,9 @@ def draw_nodes(
     node_color : string
         Color of the nodes.
     cmap : string
-        The matplotlib color map to be used for plotting. Refer to matplotlib
-        documentation for possible inputs.
+        The matplotlib color map to be used for plotting. Refer to matplotlib documentation for possible inputs.
     node_shape : string
-        Shape of the nodes. Refer to matplotlib documentation for available 
-        marker types.
+        Shape of the nodes. Refer to matplotlib documentation for available marker types.
     edge_colors : string
         Color of the node borders.
     line_widths : integer
@@ -58,6 +57,7 @@ def draw_nodes(
     label : string
         Matplotlib label of plotting instance.
     """
+    
     model = self.model
     if parameter_results is None:
         parameter_results = []
@@ -177,6 +177,7 @@ def draw_links(
     link_arrows=False
 ):
     """Draws continuous link data onto the figure.
+    
     Arguments
     ---------
     ax : axes._subplots.AxesSubplot
@@ -188,8 +189,7 @@ def draw_links(
     edge_colors : string
         Color of links.
     cmap : string
-        The matplotlib color map to be used for plotting. Refer to matplotlib
-        documentation for possible inputs.
+        The matplotlib color map to be used for plotting. Refer to matplotlib documentation for possible inputs.
     widths : integer, array-like
         Integer representing all link widrths, or array of widths for each link.
     vmin : integer
@@ -197,8 +197,7 @@ def draw_links(
     vmax : integer
         The maximum value of the color bar.
     link_style : string
-        The style (solid, dashed, dotted, etc.) of the links. Refer to 
-        matplotlib documentation for available line styles.
+        The style (solid, dashed, dotted, etc.) of the links. Refer to matplotlib documentation for available line styles.
     link_arrows : boolean
         Determines if an arrow is drawn in the direction of flow of the pump.
     """
@@ -357,16 +356,15 @@ def draw_base_elements(
     """
     Draws base elements (nodes, links, reservoirs, tanks, pumps, and valves)
     without any data associated with the elements.
+    
     Arguments
     ---------
     ax : axes._subplots.AxesSubplot
         Matplotlib axes object.
     nodes : boolean
-        Determines if base nodes with no data associated with them are drawn.
-        Set to False for all functions excep plot_basic_elements by default.
+        Determines if base nodes with no data associated with them are drawn. Set to False for all functions excep plot_basic_elements by default.
     links : boolean
-        Determines if base links with no data associated with them are drawn.
-        Set to False for all functions that deal with link data plotting.
+        Determines if base links with no data associated with them are drawn. Set to False for all functions that deal with link data plotting.
     reservoirs : boolean
         Determines if reservoirs with no data associated with them are drawn.
     tanks : boolean
@@ -380,11 +378,9 @@ def draw_base_elements(
     reservoir_size : integer
         The size of the reservoir marker on the plot in points^2. 
     reservoir_color : string
-        The color of the reservoir marker. Refer to matplotlib documentation for 
-        available colors.
+        The color of the reservoir marker. Refer to matplotlib documentation for available colors.
     reservoir_shape : string
-        The shape of the reservoir marker. Refer to matplotlib documentation for
-        available marker types.
+        The shape of the reservoir marker. Refer to matplotlib documentation for available marker types.
     reservoir_border_color : string
         The color of the border around the reservoir marker.
     reservoir_border_width : integer
[…]
