+++ viswaternet/drawing/continuous.py
@@ -2,6 +2,7 @@
 """
 The viswaternet.drawing.continuous module handles everything related to continuous data drawing.
 """
+
 import matplotlib as mpl
 from viswaternet.network import processing
 from viswaternet.utils import save_fig, normalize_parameter, unit_conversion, fancyarrowpatch_to_linecollection, label_generator
@@ -9,7 +10,6 @@ from viswaternet.drawing import base
 
 
 default_cmap = 'autumn_r'
-
 
 def plot_continuous_nodes(
     self,
@@ -79,8 +79,8 @@ def plot_continuous_nodes(
     color_bar_width=0.03,
     color_bar_height=0.8
 ):
-    """User-level function that draws continuous nodal data, base elements,
-    legends, and saves the figure.
+    """User-level function that draws continuous nodal data, base elements, legends, and saves the figure.
+    
     Arguments
     ---------
     ax : axes._subplots.AxesSubplot
@@ -133,8 +133,7 @@ def plot_continuous_nodes(
     valves : boolean
         Determines if valves with no data associated with them are drawn.
     cmap : string
-        The matplotlib color map to be used for plotting. Refer to matplotlib
-        documentation for possible inputs.
+        The matplotlib color map to be used for plotting. Refer to matplotlib documentation for possible inputs.
     color_bar_title : string
         The title of the color bar.
     node_size : integer, array-like
@@ -144,8 +143,7 @@ def plot_continuous_nodes(
     max_size : integer
         Maximum size of nodes to be used with normalize_parameter.
     node_shape : string
-        Shape of the nodes. Refer to matplotlib documentation for available 
-        marker types.
+        Shape of the nodes. Refer to matplotlib documentation for available marker types.
     edge_colors : string
         Color of the node borders.
     line_widths : integer
@@ -168,20 +166,17 @@ def plot_continuous_nodes(
     legend : boolean
         Determines if the base elements legend will be drawn. 
     legend_loc : string
-        The location of the base elements legend on the figure. Refer to matplotlib
-        documentation for possible inputs.
+        The location of the base elements legend on the figure. Refer to matplotlib documentation for possible inputs.
     font_size : integer
         The font size of the non-title text for legends. 
     font_color : string
-        The color of the legend text. Refer to matplotlib documentation for 
-        available colors.
+        The color of the legend text. Refer to matplotlib documentation for available colors.
     legend_title_font_size : integer
         The font size of the title text for legends.
     draw_frame : boolean
         Determines if the frame around the legend is drawn.
     legend_sig_figs : integer
-        The number of significant figures, or decimal points, that numbers in the
-        legend will be displayed with. 0 should be passed for whole numbers.
+        The number of significant figures, or decimal points, that numbers in the legend will be displayed with. 0 should be passed for whole numbers.
     element_size_intervals : integer
         The number of intervals to be used if an element size legend is used.
     element_size_legend_title : string
@@ -195,8 +190,7 @@ def plot_continuous_nodes(
     reservoir_color : string
         The color of the reservoir marker.
     reservoir_shape : string
-        The shape of the reservoir marker. Refer to matplotlib documentation for
-        available marker types.
+        The shape of the reservoir marker. Refer to matplotlib documentation for available marker types.
     reservoir_border_color : string
         The color of the border around the reservoir marker.
     reservoir_border_width : integer
@@ -226,8 +220,7 @@ def plot_continuous_nodes(
     pump_width : integer
         The width of the pump line in points.
     pump_line_style : string
-        The style (solid, dashed, dotted, etc.) of the pump line. Refer to 
-        matplotlib documentation for available line styles.
+        The style (solid, dashed, dotted, etc.) of the pump line. Refer to matplotlib documentation for available line styles.
     pump_arrows : boolean
[…]
