+++ viswaternet/drawing/unique.py
@@ -2,6 +2,7 @@
 """
 The viswaternet.drawing.unique module handles custom data, excel data, and unique data drawing.
 """
+
 import numpy as np
 import pandas as pd
 import matplotlib as mpl
@@ -99,27 +100,18 @@ def plot_unique_data(
     color_bar_width=0.03,
     color_bar_height=0.8
 ):
-    """A complex function that accomplishes tasks relating to categorical data,
-    or 'unique' as used in viswaternet, as well as data not retrieved from WNTR.
+    """A complex function that accomplishes tasks relating to categorical data, or 'unique' as used in viswaternet, as well as data not retrieved from WNTR.
 
-    There are three distinct modes of operation, and which one is used is controlled
-    by the 'parameter' argument, which differs from previous use of the argument.
+    There are three distinct modes of operation, and which one is used is controlled by the 'parameter' argument, which differs from previous use of the argument.
 
-    Setting the parameter argument to 'demand_patterns', 'diameter', 
-    or 'roughness' simply plots that parameter. These parameters are treated 
-    differently from others because in usually they are categorical. For 
-    instance, pipe diameters are not randomly chosen, and instead are chosen from
-    a list of standard pipe sizes.
+    Setting the parameter argument to 'demand_patterns', 'diameter', or 'roughness' simply plots that parameter. These parameters are treated differently from others because in usually they are categorical. For 
+    instance, pipe diameters are not randomly chosen, and instead are chosen from a list of standard pipe sizes.
 
-    When the parameter argument is set to 'excel_data', the function deals
-    with excel data, or data imported from an .xlsx file. Two excel columns 
-    with elements and data pairs are provided by the user, which are then 
+    When the parameter argument is set to 'excel_data', the function deals with excel data, or data imported from an .xlsx file. Two excel columns with elements and data pairs are provided by the user, which are then 
     converted into a format usable by viswaternet for plotting. 
 
-    When the parameter argument is set to 'custom_data', the function deals with
-    data directly inside of python. The user should expect to format the data
-    themselves, although this shouldn't be difficult. An example of 'custom_data'
-    being used can be seen in example 5 located in the github repository. 
+    When the parameter argument is set to 'custom_data', the function deals with data directly inside of python. The user should expect to format the data themselves, although this shouldn't be difficult. An example of 'custom_data'
+    being used can be seen in example 10 located in the github repository. 
 
     Arguments
     ---------
@@ -145,24 +137,19 @@ def plot_unique_data(
         Column A in excel is considered the 0th column for use with viswaternet.
 
     custom_data_values : array-like
-        Similar to 'excel_columns' two values should be provided. The first value
-        should be an array with element names, and the second should be one with
-        the element data.
+        Similar to 'excel_columns' two values should be provided. The first value should be an array with element names, and the second should be one with the element data.
 
     unit : string
         The unit that the network data is to be converted to.
 
     intervals : integer, string
-        If set to 'automatic' then intervals are created automatically on a 
-        equal interval basis. Otherwise, it is the edges of the intervals to be
-        created. intervals array length should be num_intervals + 1.
+        If set to 'automatic' then intervals are created automatically on a equal interval basis. Otherwise, it is the edges of the intervals to be created. intervals array length should be num_intervals + 1.
 
     interval_node_size_list : integer, array-like
         List of node sizes for each interval.
 
     interval_node_shape_list : string, array-like
-        List of node shapes for each interval. Refer to matplotlib documentation
-        for available marker types.
+        List of node shapes for each interval. Refer to matplotlib documentation for available marker types.
 
[…]
