### 2023-03-15 47b7d70
modify table inputs
