+++ viswaternet/drawing/continuous.py
@@ -100,10 +100,13 @@ def plot_continuous_nodes(
         - leak_area
         - leak_discharg_coeff
         - quality
+        
     value : integer, string
         For time-varying parameters only. Specifies which timestep or data
         summary will be plotted.
+        
         .. rubric:: Possible Inputs
+        
         ======================= =========================================
             int                 Plots element data for specified timestep
             'min'               Plots minimum data point for each element
@@ -112,6 +115,7 @@ def plot_continuous_nodes(
             'stddev'            Plots standard deviation for each element
             'range'             Plots range for each element
         ======================= =========================================
+        
     unit : string
         The unit that the network data is to be converted to.
     vmin : integer
@@ -438,10 +442,13 @@ def plot_continuous_links(
         - friction_factor
         - reaction_rate
         - quality
+        
     value : integer, string
         For time-varying parameters only. Specifies which timestep or data
         summary will be plotted.
+        
         .. rubric:: Possible Inputs
+        
         ======================= =========================================
             int                 Plots element data for specified timestep
             'min'               Plots minimum data point for each element
@@ -450,6 +457,7 @@ def plot_continuous_links(
             'stddev'            Plots standard deviation for each element
             'range'             Plots range for each element
         ======================= =========================================
+        
     unit : string
         The unit that the network data is to be converted to.
     widths : integer, array-like
