### 2023-03-15 37f7ba2
Update AUTHORS.rst
