+++ setup.py
@@ -12,8 +12,8 @@ with open('HISTORY.rst') as history_file:
 
 requirements = ['pandas',
                'numpy',
-               'matplotlib>=3.3.0',
-               'wntr>= 0.5.0',
+               'matplotlib>=3.5.0',
+               'wntr< 0.5.0',
                'imageio',
                'networkx>=2.7']
 
@@ -22,7 +22,7 @@ test_requirements = [ ]
 setup(
     author="Tyler Trimble",
     author_email='TylerL.Trimble@gmail.com',
-    python_requires='3.8, 3.9, 3.10',
+    python_requires='3.8, 3.9',
     classifiers=[
         'Development Status :: 2 - Pre-Alpha',
         'Intended Audience :: Developers',
@@ -30,8 +30,7 @@ setup(
         'Natural Language :: English',
         'Programming Language :: Python :: 3.8',
         'Programming Language :: Python :: 3.9',
-        'Programming Language :: Python :: 3.10',
-    ],
+            ],
     description="A python package for easy generation and customization of water network graphs.",
     install_requires=requirements,
     license="MIT license",
