### 2023-03-15 tylertrimble/viswaternet#101 @meghnathomas
Hi @thurber thank you for your patience! I think we have addressed all of your comments (please see below). Please note that we haven't yet released a new PyPI version or drafted a release with the changes.

**[General]**
1. > Can you comment on the contributions of the paper co-author Lina Sela? I only ask because they don't show up in the commit history, which is fine, just need to clarify.

We've updated the [author contributions, CRediT-style,](https://viswaternet.readthedocs.io/en/latest/authors.html) in the docs.

2. > Can you add a section to the README.md or a contributing.md file outlining your community guidelines (i.e. how folks can/should contribute to the code). See GitHub's recommendations [here](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors).

The [Contributing](https://viswaternet.readthedocs.io/en/latest/contributing.html) page in the docs has been updated, and we've included a link to it in the README.

3. > Can you add a section to the README.md that says how to run the automated tests

We've added a [small Testing section](https://viswaternet.readthedocs.io/en/latest/contributing.html#testing) to the Contributing page, and link to it in the README.

4. > There are a few deprecation warnings generated when running the tests; in particular the matplotlib ones may be cause for concern -- you don't want upstream changes to matplotlib to break your code. It would be preferable to update your use of matplotlib APIs to the latest syntax, but you could alternatively pin the maximum allowed matplotlib version in your requirements.txt.

We examined this in #103 and changed other dependency version requirements to fix this issue.

**[Docs/Code]**
1. > In the GitHub README.md example (also in the Getting Started docs section), in case the user is not running in a notebook it would be helpful to explicitly mention the need to run plt.show() to see the plot.

In the [Getting Started](https://viswaternet.readthedocs.io/en/latest/gettingstarted.html) section of the README and the docs, we now recommend using plt.show() to display the plot if the user's  IDE does not support interactive plotting (e.g., IDLE) or interactive mode is off.

2. > When running through the example in the GitHub README.md, a warning is generated -- would be good to either suppress it if it's irrelevant to your package, or more fully explain it so the user knows what should be done: `/Users/thur961/viswaternet/venv/lib/python3.10/site-packages/wntr/epanet/io.py:2082: UserWarning: Not all curves were used in "/Users/thur961/viswaternet/CTown.inp"; added with type None, units conversion left to user warnings.warn('Not all curves were used in "{}"; added with type None, units conversion left to user'.format(self.wn.name))`

This error indicates that the .INP file that contains the network topology includes curves that are not applied to any of the elements in the network and is not fatal. It is generated by the package WNTR that we rely on to extract network characteristics and run simulations. We decided not to suppress any WNTR warnings so we now include a [Warning Messages](https://viswaternet.readthedocs.io/en/latest/contributing.html#warning-messages) section on the Contributing page where we will continue to add warnings that appear to us or other users.

3. > In Example 2, when plotting the pumps as dotted lines, it would be nice if the legend also updated to show pumps in dotted lines. Furthermore, it would be good to point out that the arguments are following the matplotlib notation for colors and line types, etc., and perhaps link to the matplotlib docs page that describes them.

In #104 we raised and fixed the issue of accurately representing pumps and pipes in the legend. We have also added links to the Getting Started sections of the docs and README directing users towards relevant matplotlib references.

4. > For all plots, it would be good to show the units along with the legend, and to perhaps somehow annotate the chosen variable and timestep or temporal aggregation method (I see that user can input a custom legend title as one option; if that's preferred then would help to mention it in the docs).

We raised and fixed this in Issue #105. By default, continuous and discrete plots now include legend titles describing the plotting parameter (e.g., "pressure"), value (e.g., "at timestep 10" or "mean"), and units.

5. > Could be nice to support an `ax=None` default argument that creates a matplotlib fig/ax on behalf of the user. This would follow the precedent of libraries like `pandas/xarray`.

We liked this idea! It is now optional for the user to provide the `ax` argument (#106). The docs and examples now reflect this change.

We really appreciate your thoughtful feedback and suggestions! I hope we have sufficiently addressed your comments.
