### 2023-03-15 tylertrimble/viswaternet#111 @meghnathomas
Hi @erexer thank you for your patience! I think we have addressed all of your comments here. Please note that we haven't yet released a new PyPI version or drafted a release with the changes.

**Contribution**
1. > how was tylertrimble involved? The only info I have is the commit history so I want to double check that the authors are "appropriate and complete"

We have updated the [author contributions, CRediT-style,](https://viswaternet.readthedocs.io/en/latest/authors.html) in the docs.

**Installation**

2. > since you list as working with python 3.8+, it should work for 3.11.0. Let me know if you’re able to reproduce this, if not I can give a better trace

Thank you for pointing this out! We have addressed this comment in #109. We use the package WNTR to extract water network characteristics and run simulations. WNTR does not yet support Python 3.11 (and further testing revealed that WNTR is [yet to make changes to its setup file that fully support Python 3.10](https://github.com/USEPA/WNTR/issues/331)) so we have decided to only test VisWaterNet on Python version 3.8 and 3.9 for now. We will add support for Python 3.10 and 3.11 after WNTR makes its own updates.

3. > python 3.10.9 python -m unittest throws a lot of warnings
`MatplotlibDeprecationWarning: The get_cmap function was deprecated in Matplotlib 3.7 and will be removed two minor releases later. Use ``matplotlib.colormaps[name]`` or ``matplotlib.colormaps.get_cmap(obj)`` instead. cmap = mpl.cm.get_cmap(cmap)`

We examined this in https://github.com/tylertrimble/viswaternet/issues/103 and changed other dependency version requirements to fix this issue.

**Examples**

4. > They look great! I’d recommend standardizing the example numbering for the python scripts and the examples in the docs. example2_network_diameters.py would be named example8_network_diameters.py and produce identical output to [Example 8](https://viswaternet.readthedocs.io/en/latest/examples.html#example-8-categorical-data-plot-for-link-diameter) in the readthedocs. (personal preference so feel free to ignore)

This is a good idea! We updated the numbering convention of the example files to mirror the order they appear in in the docs.

5. > Getting the same error from every example:
`UserWarning: Not all curves were used in "Networks/CTown.inp"; added with type None, units conversion left to user warnings.warn('Not all curves were used in "{}"; added with type None, units conversion left to user'.format([self.wn.name](http://self.wn.name/)))`

This error indicates that the .INP file that contains the network topology includes curves that are not applied to any of the elements in the network. It is generated by the package WNTR that we rely on to extract network characteristics and run simulations. We decided not to suppress any WNTR warnings so we now include a [Warning Messages](https://viswaternet.readthedocs.io/en/latest/contributing.html#warning-messages) section on the Contributing page where we will continue to add warnings that appear to us or other users. We have also modified the .INP file of the CTown network so this warning should no longer pop up.

**Tests**

6. > python 3.10.9 throws warning from python -m unittest
`MatplotlibDeprecationWarning: Passing *transOffset* without *offsets* has no effect. This behavior is deprecated since 3.5 and in 3.6, *transOffset* will begin having an effect regardless of *offsets*. In the meantime, if you wish to set *transOffset*, call collection.set_offset_transform(transOffset) explicitly.`

This issue was also fixed in https://github.com/tylertrimble/viswaternet/issues/103 when we changed other dependency version requirements.

**Documentation**

7. > would recommend adding the documentation link in the GitHub sidebar
also recommend running it through a spell checker since there were a few spelling mistakes

We have added the link to the docs in the sidebar and have checked spellings in the text.

**Paper**

8. > Does “extracts WDS characteristics and simulation results of a user-provided .INP file” mean something different than “1. Import and visualize custom data generated either in Python scripts or in Excel files.” are the Python scripts or Excel files generating anything other than .INP files?

Yes. The .INP file contains all properties of the WDS network model (such as pipe diameters, nodal elevations, demand patterns, node coordinates, etc.) whereas "custom data" refers to data points that correspond to either nodes or links. We use the package WNTR to extract the network topology from the .INP file (that we use to set up the network layout) and to perform steady-state hydraulic simulations of the network (the results of which we use to plot discrete/continuous plots of parameters such as pressure, flow rate, demand, etc.). We allow users to plot custom data that they may generate through analyses of their own by allowing them to import extra information through other Excel files or other Python scripts. Examples [9](https://viswaternet.readthedocs.io/en/latest/examples.html#example-9-importing-and-plotting-categorical-data-from-an-excel-file) and [10](https://viswaternet.readthedocs.io/en/latest/examples.html#example-10-plotting-custom-data-generated-within-a-python-script) in the docs illustrate examples of custom data plotting.

9. > recommend changing “Add a second dimension of information to a plot by customizing node markers and sizes and link styles and widths.” to “Add … customizing node markers/sizes and link styles/widths.” for clarity.

This has been fixed.

10. > “Convert units (e.g., a user can choose to represent length data in units of m, cm, ft, or in.” missing closing parenthesis.

This has been fixed.

We thank you again for your time and effort invested in reviewing this package! I hope we have sufficiently addressed your comments.
