+++ setup.py
@@ -22,9 +22,9 @@ test_requirements = [ ]
 setup(
     author="Tyler Trimble",
     author_email='TylerL.Trimble@gmail.com',
-    python_requires='3.8, 3.9',
+    python_requires='>=3.8, <=3.9.16',
     classifiers=[
-        'Development Status :: 2 - Pre-Alpha',
+        'Development Status :: 5 - Production/Stable',
         'Intended Audience :: Developers',
         'License :: OSI Approved :: MIT License',
         'Natural Language :: English',
@@ -42,6 +42,6 @@ setup(
     test_suite='tests',
     tests_require=test_requirements,
     url='https://github.com/tylertrimble/viswaternet',
-    version='1.0.0',
+    version='1.1.0',
     zip_safe=False,
 )
