### 2023-03-17 tylertrimble/viswaternet#111 @erexer
Looks great! Thanks for making those changes.

Couple of small things:
* [github recommends](https://github.com/github/renaming) renaming the `master` branch to `main`
* `More examples can be found in the Examples folder.` [Examples link](https://viswaternet.readthedocs.io/en/latest/examples.html#example-11-creating-gifs) is broken in the docs (capital E vs lowercase e)
