### 2023-03-21 tylertrimble/viswaternet#111 @meghnathomas
Hi @erexer,

Thank you for the comments!

We have
1. renamed the `master` branch to `main`
2. fixed the links to the examples folder in the Getting Started and Example Applications sections of the docs.
