### 2023-03-28 tylertrimble/viswaternet#101 @thurber
Thanks @meghnathomas and team for the thorough response and updates!

A few final thoughts on paper/documentation:

- [ ] In the paper and in the docs, can you give a little more context to the .INP file format? A user could still get great benefit from your package without being aware of this format. Just linking/citing other documentation would be sufficient (i.e. to [here](https://epanet22.readthedocs.io/en/latest/back_matter.html#input-file-format)).
- [ ] State of the Field - this is typically its own section, but you mostly cover it in Statement of Need. Are there other packages out there aside from the ones you're building upon that do similar things? If so, might be worth adding a State of the Field section; otherwise I think it's good as is.
- [ ] This is just personal preference, but ending the paper with a bulleted list feels a little abrupt; consider adding a final paragraph mentioning how VisWaterNet is being used in your own or others' research/work.
- [ ] Rossman et al 2020 reference is missing a doi/url.
