+++ paper/paper.bib
@@ -1,8 +1,9 @@
-@misc{rossman2020,
+@report{rossman2020,
 title={EPANET 2.2 User Manual, Report EPA/600/R-20/133},
 author={Rossman, Lewis A and Woo, Hyoungmin and Tryby, Michael and Shang, Feng and Janke, Robert and Haxton, Terranna},
 year={2020},
-publisher={US Environmental Protection Agency, Washington, DC, USA}
+publisher={US Environmental Protection Agency, Washington, DC, USA},
+url={https://cfpub.epa.gov/si/si_public_record_Report.cfm?dirEntryId=348882&Lab=CESER}
 }
 
 @article{klise2017, 
@@ -35,11 +36,15 @@ pages={420–431}}
   doi={10.1007/978-3-642-18638-7_6}
 }
 
-@techreport{hagberg2008,
-  title={Exploring network structure, dynamics, and function using NetworkX},
-  author={Hagberg, Aric and Swart, Pieter and S Chult, Daniel},
-  year={2008},
-  institution={Los Alamos National Lab.(LANL), Los Alamos, NM (United States)}
+@article{hagberg2008,
+    title = {Exploring network structure, dynamics, and function using networkx},
+    author = {Hagberg, Aric and Swart, Pieter and S Chult, Daniel},
+    url = {https://www.osti.gov/biblio/960616}, journal = {},
+    number = ,
+    volume = ,
+    place = {United States},
+    year = {2008},
+    month = {1}
 }
 
 @article{hunter2007,
@@ -64,4 +69,24 @@ pages={420–431}}
   year={2012},
   publisher={American Society of Civil Engineers},
   doi = {10.1061/(ASCE)WR.1943-5452.0000191}
+}
+
+@article{arandia2018r,
+  title={An R package for EPANET simulations},
+  author={Arandia, Ernesto and Eck, Bradley J},
+  journal={Environmental Modelling \& Software},
+  volume={107},
+  pages={59--63},
+  year={2018},
+  publisher={Elsevier},
+  doi = {10.1016/j.envsoft.2018.05.016}
+}
+
+@inproceedings{eliades2016epanet,
+  title={EPANET-MATLAB toolkit: An open-source software for interfacing EPANET with MATLAB},
+  author={Eliades, Demetrios G and Kyriakou, Marios and Vrachimis, Stelios and Polycarpou, Marios M},
+  booktitle={Proc. 14th International Conference on Computing and Control for the Water Industry (CCWI)},
+  volume={8},
+  year={2016},
+  doi = {10.5281/zenodo.831493}
 }
