+++ paper/paper.bib
@@ -71,7 +71,7 @@ pages={420–431}}
   doi = {10.1061/(ASCE)WR.1943-5452.0000191}
 }
 
-@article{arandia2018r,
+@article{arandia2018,
   title={An R package for EPANET simulations},
   author={Arandia, Ernesto and Eck, Bradley J},
   journal={Environmental Modelling \& Software},
@@ -82,7 +82,7 @@ pages={420–431}}
   doi = {10.1016/j.envsoft.2018.05.016}
 }
 
-@inproceedings{eliades2016epanet,
+@inproceedings{eliades2016,
   title={EPANET-MATLAB toolkit: An open-source software for interfacing EPANET with MATLAB},
   author={Eliades, Demetrios G and Kyriakou, Marios and Vrachimis, Stelios and Polycarpou, Marios M},
   booktitle={Proc. 14th International Conference on Computing and Control for the Water Industry (CCWI)},
