+++ paper/paper.md
@@ -28,11 +28,11 @@ Water distribution systems (WDSs) are complex networks and are typically represe
 
 # Statement of need
 
-Existing open-source tools for analyzing WDSs, such as `EPANET` [@rossman2020] and the Water Network Tool for Resilience (`WNTR`) [@klise2017], allow users to perform extended period simulations and analyze system hydraulic performance under a range of different conditions. However, while these tools provide basic network visualization options, they are limited in their flexibility and capabilities. On the other hand, popular software such as `Cytoscape` [@shannon2003] and `Graphviz` [@ellson2004] allow the easy integration of attribute data into sophisticated network plots, but it is challenging to apply them to provide meaningful results in the context of WDSs. `VisWaterNet`  was developed to provide an accessible and multifunctional approach to WDS-specific network visualization. The package is accompanied by clear documentation and code examples in order to facilitate ease of adoption, and can prove to be a helpful tool for researchers and practitioners in the domain of WDS analysis. Users are welcome to further develop, improve and extend these open-source scripts.
+Existing open-source tools for analyzing WDSs, such as `EPANET` [@rossman2020] and the Water Network Tool for Resilience Python package (`WNTR`) [@klise2017], allow users to perform extended period simulations and analyze system hydraulic performance under a range of different conditions. However, while these tools provide basic network visualization options, they are limited in their flexibility and capabilities. `EPANET` wrappers in other programming languages also offer few plotting functions; for example, an `EPANET`-interfacing toolkit developed for MATLAB [@eliades2016] only generates network layouts, and a similar R application [arandia2018] generates basic network representations. On the other hand, popular software such as `Cytoscape` [@shannon2003] and `Graphviz` [@ellson2004] allow the easy integration of attribute data into sophisticated network plots, but it is challenging to apply them to provide meaningful results in the context of WDSs. `VisWaterNet`  was developed to provide an accessible and multifunctional approach to WDS-specific network visualization. The package is accompanied by clear documentation and code examples in order to facilitate ease of adoption, and can prove to be a helpful tool for researchers and practitioners in the domain of WDS analysis. Users are welcome to further develop, improve and extend these open-source scripts.
 
 # Functionality
 
-`VisWaterNet` extracts WDS characteristics and simulation results of a user-provided .INP file using `WNTR` [@klise2017], and then constructs a `NetworkX` [@hagberg2008] graph on a `Matplotlib` [@hunter2007] figure axis. The hydraulic simulation results are based on the `EPANET 2.2` [@rossman2020] hydraulic engine. The main plotting functions offered by `VisWaterNet` include:
+`VisWaterNet` extracts WDS characteristics and simulation results from a standardized user-provided input file, i.e., `EPANET` .INP file [@rossman2020], using `WNTR` [@klise2017], and then constructs a `NetworkX` [@hagberg2008] graph on a `Matplotlib` [@hunter2007] figure axis. The hydraulic simulation results are based on the `EPANET 2.2` [@rossman2020] hydraulic engine. The main plotting functions offered by `VisWaterNet` include:
 
 1.	Continuous plotting - users can choose to represent numerical data in a *continuous* manner, i.e., by assigning  nodes or links colors chosen from a color gradient scale. \autoref{fig:fig1} (left) shows an example of a continuous plot of mean nodal pressure data for network CTown [@ostfeld2012] generated by `VisWaterNet`.
 
@@ -53,6 +53,8 @@ Additionally, `VisWaterNet` can perform the following functions:
 5.	Customize the number of intervals or width of intervals for discrete data plots, and hide intervals with no data points.
 6.	Customize the appearance of specific nodes and links.
 7.	Add text labels to specific nodes and links.
+
[…]
