### 2023-04-01 tylertrimble/viswaternet#101 @meghnathomas
Hi @thurber thank you for these comments on the paper and docs!

We have added a short description of a .INP file to the paper and provided a link to the EPANET docs section you suggested in the Overview and Getting Started sections of the docs.
We added a line in the Statement of Need section about packages in other programming languages that offer basic visualization options.
We added a short concluding paragraph to the paper.
And we added a URL to the Rossman reference.

Please let us know if you think we should make any more changes!
