### 2023-04-02 tylertrimble/viswaternet#101 @thurber
Thanks @meghnathomas, nice work! I will recommend the paper be accepted.
