### 2023-04-03 308c412
Change auto figure generation
Figures and axes are generated when a function is called and not when the model is initialized, this solves a number of issues.
