### 2023-04-03 bcf770b
Overhaul animate_plot
For the 5th time.... to allow for unique data animation and more than 256 colors (GIF to mp4)
