+++ viswaternet/drawing/animate.py
@@ -1,25 +1,23 @@
 # -*- coding: utf-8 -*-
 """
-The viswaternet.drawing.animate module facilitates the creation of .gif files 
+The viswaternet.drawing.animate module facilitates the creation of .gif files
 animating data across timesteps.
 """
 import numpy as np
 import matplotlib.pyplot as plt
 import imageio
 from viswaternet.network import processing
-from viswaternet.utils import unit_conversion
+from viswaternet.utils import unit_conversion,convert_excel
 
 
 def animate_plot(
     self,
-    ax,
     function,
-    data_type,
-    parameter_type,
+    ax=None,
     fps=3,
     first_timestep=0,
     last_timestep=None,
-    gif_save_name="gif",
+    gif_save_name="animation",
     unit='s',
     **kwargs
 ):
@@ -77,67 +75,91 @@ def animate_plot(
         can be passed into animate_plot.
     """
     model = self.model
-    frames=[]
-    timesteps = int(
-        model["wn"].options.time.duration /
-        model["wn"].options.time.report_timestep
-    )
-    values = range(timesteps)
-    if last_timestep is not None:
-        values = values[first_timestep:last_timestep]
-
-    if data_type == "continuous":
+    if ax is None:
+        if ax is None:
+            fig, ax = plt.subplots(figsize=self.figsize)  
+            ax.set_frame_on(self.axis_frame)
+    frames = []
+    if function == self.plot_unique_data:
+        timesteps = last_timestep
+        
+        values = range(timesteps)
+        if last_timestep is not None:
+            values = values[first_timestep:last_timestep]
+    else:
+        timesteps = int(
+            model["wn"].options.time.duration /
+            model["wn"].options.time.report_timestep
+        )
+        values = range(timesteps)
+        if last_timestep is not None:
+            values = values[first_timestep:last_timestep]
+    
+    if function == self.plot_unique_data:
+        parameter_type = kwargs.get("parameter_type",None)
+        data_type = kwargs.get("data_type",None)
+        try:
+            custom_data_values = kwargs.get("custom_data_values")
+            data_values = custom_data_values[1]
+            
+        except KeyError:
+            excel_columns = kwargs.get("excel_columns",None)
+            data_values = []
+            for i in excel_columns[1]:
+                data = convert_excel(
+                    self, kwargs.get("parameter",None), parameter_type, data_type, excel_columns[0], excel_columns[1]
+                )
+                data_values.append(data)
+                
+        if data_type == "continuous":
+            if kwargs.get("vmin", None) is None or kwargs.get("vmax", None) is None:
+                kwargs["vmin"],kwargs["vmax"] = make_vmin_vmax(data_values,kwargs)
+                    
+        if data_type == 'discrete':
+            kwargs["disable_interval_deleting"] = True
+            
+            if kwargs.get("intervals", None) is None:
+                kwargs["intervals"] = make_intervals(data_values,kwargs)
+    
+    if function == self.plot_continuous_nodes or function == self.plot_discrete_nodes:
+        parameter_type = 'node'
+        
+    elif function == self.plot_continuous_links or function == self.plot_discrete_links:
+        parameter_type = 'link'
+        
+    if function == self.plot_continuous_nodes or function == self.plot_continuous_links:
         if kwargs.get("vmin", None) is None or kwargs.get("vmax", None) is None:
-            if parameter_type == "link":
-                parameter_results, link_list = processing.get_parameter(
-                    self, "link", kwargs.get(
-                        "parameter"), kwargs.get("value", None)
-                )
-            if parameter_type == "node":
-                parameter_results, node_list = processing.get_parameter(
-                    self, "node", kwargs.get(
[…]
