### 2023-04-04 44a1d3d
Remove removed parameters from documentation
