### 2023-04-04 d9fb70a
Make it so savefig is set to False by default
