+++ viswaternet/drawing/discrete.py
@@ -392,7 +392,7 @@ def plot_discrete_nodes(
     interval_label_list=None,
     interval_node_border_color_list=None,
     interval_node_border_width_list=None,
-    savefig=True,
+    savefig=False,
     save_name=None,
     dpi='figure',
     save_format='png',
@@ -797,7 +797,7 @@ def plot_discrete_links(
     legend_title=None,
     legend_loc_1="upper right",
     legend_loc_2="lower right",
-    savefig=True,
+    savefig=False,
     save_name=None,
     dpi='figure',
     save_format='png',
