+++ viswaternet/drawing/continuous.py
@@ -37,7 +37,7 @@ def plot_continuous_nodes(
     line_widths=None,
     legend=True,
     legend_loc="upper right",
-    savefig=True,
+    savefig=False,
     save_name=None,
     dpi='figure',
     save_format='png',
@@ -378,7 +378,7 @@ def plot_continuous_links(
     color_bar_title=None,
     legend=True,
     legend_loc="upper right",
-    savefig=True,
+    savefig=False,
     save_name=None,
     dpi='figure',
     save_format='png',
