### 2023-04-04 88685b2
Check for division by zero in parameter normalization
