### 2023-04-04 05a81d5
Add timestep # detection for unique data plotting
