+++ paper/paper.md
@@ -28,7 +28,7 @@ Water distribution systems (WDSs) are complex networks and are typically represe
 
 # Statement of need
 
-Existing open-source tools for analyzing WDSs, such as `EPANET` [@rossman2020] and the Water Network Tool for Resilience Python package (`WNTR`) [@klise2017], allow users to perform extended period simulations and analyze system hydraulic performance under a range of different conditions. However, while these tools provide basic network visualization options, they are limited in their flexibility and capabilities. `EPANET` wrappers in other programming languages also offer few plotting functions; for example, an `EPANET`-interfacing toolkit developed for MATLAB [@eliades2016] only generates network layouts, and a similar R application [arandia2018] generates basic network representations. On the other hand, popular software such as `Cytoscape` [@shannon2003] and `Graphviz` [@ellson2004] allow the easy integration of attribute data into sophisticated network plots, but it is challenging to apply them to provide meaningful results in the context of WDSs. `VisWaterNet`  was developed to provide an accessible and multifunctional approach to WDS-specific network visualization. The package is accompanied by clear documentation and code examples in order to facilitate ease of adoption, and can prove to be a helpful tool for researchers and practitioners in the domain of WDS analysis. Users are welcome to further develop, improve and extend these open-source scripts.
+Existing open-source tools for analyzing WDSs, such as `EPANET` [@rossman2020] and the Water Network Tool for Resilience Python package (`WNTR`) [@klise2017], allow users to perform extended period simulations and analyze system hydraulic performance under a range of different conditions. However, while these tools provide basic network visualization options, they are limited in their flexibility and capabilities. `EPANET` wrappers in other programming languages also offer few plotting functions; for example, an `EPANET`-interfacing toolkit developed for MATLAB [@eliades2016] only generates network layouts, and a similar R application [@arandia2018] generates basic network representations. On the other hand, popular software such as `Cytoscape` [@shannon2003] and `Graphviz` [@ellson2004] allow the easy integration of attribute data into sophisticated network plots, but it is challenging to apply them to provide meaningful results in the context of WDSs. `VisWaterNet`  was developed to provide an accessible and multifunctional approach to WDS-specific network visualization. The package is accompanied by clear documentation and code examples in order to facilitate ease of adoption, and can prove to be a helpful tool for researchers and practitioners in the domain of WDS analysis. Users are welcome to further develop, improve and extend these open-source scripts.
 
 # Functionality
 
