+++ paper/paper.bib
@@ -39,9 +39,7 @@ pages={420–431}}
 @article{hagberg2008,
     title = {Exploring network structure, dynamics, and function using networkx},
     author = {Hagberg, Aric and Swart, Pieter and S Chult, Daniel},
-    url = {https://www.osti.gov/biblio/960616}, journal = {},
-    number = ,
-    volume = ,
+    url = {https://www.osti.gov/biblio/960616},
     place = {United States},
     year = {2008},
     month = {1}
