### 2023-04-04 #5139 @meghnathomas
Hi @crvernon thank you for pointing these out!

1. The arandia2018 reference has been modified in the text
2. The hagberg2008 bibtex reference has been modified according to your suggestion
